IF-19: Refactorings
