+++ fusionfrontend/src/app/components/ui/item-options-menu/item-options-menu.component.ts
@@ -33,6 +33,10 @@ export class ItemOptionsMenuComponent implements OnInit {
   }
 
   ngOnInit(): void {
+   this.initMenuItems();
+  }
+
+  private initMenuItems() {
     switch (this.type) {
       case ItemOptionsMenuType.DELETE:
         this.menuActions = [
