+++ fusionfrontend/src/app/ecosystem/components/content/asset-type-template/asset-type-template-field-row/asset-type-template-field-row.component.spec.ts
@@ -14,8 +14,7 @@
  */
 
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-
-import { AssetTypeTemplateFieldRowComponent } from './asset-type-template-field-row.component';
+import { AssetTypeTemplateFieldRowComponent } from '../../asset-type-template-field-row/asset-type-template-field-row.component';
 
 describe('AssetTypeTemplateMetricRowComponent', () => {
   let component: AssetTypeTemplateFieldRowComponent;
